backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.8.6
asttokens==2.1.0
autopep8==2.0.0
billiard==3.6.4.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
commonmark==0.9.1
coverage==6.5.0
Django==4.1.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-rest-framework==0.1.0
django-silk==5.0.2
django-timezone-field==5.0
djangorestframework==3.14.0
dparse==0.6.2
gevent==22.10.2
gprof2dot==2022.7.29
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
isort==5.10.1
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.8.0
logilab-astng==0.24.3
logilab-common==1.9.7
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
pika==1.3.1
platformdirs==2.5.4
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
pycodestyle==2.9.1
Pygments==2.13.0
pylint==2.11.1
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
rich==12.6.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zope.event==4.5.0
zope.interface==5.5.2
